back/requirements.txt
django==3.1
djangorestframework==3.11.0
django-rest-knox==4.1.0
django-cors-headers==3.4.0
celery==4.4.7
psycopg2==2.8.5
chardet==3.0.4
langdetect==1.0.8
requests==2.24.0
